Use limitSelect instead of TwoPane
TwoPane *always* shows the first two windows when the master window is
focused. limitSelect instead keeps showing the last shown slave window
when the master is focused. Sometimes the TwoPane behaviour is more
desirable but limitSelect is able to increase the number of shown
windows and is more versatile, so it was used instead of TwoPane.